Scripts/requirements.txt
matplotlib==3.3.3
numpy==1.19.4
pyserial==3.4
wavio==0.0.4